Update iF2RDF.py
Replaced the hashlib sha1 code with UUID sha5 code. This line changes the TFL Naptan Atco code to the correct SHA1 UUID [line 101]. The code requires the UUID package to work [line 12].  Note that the result is not all unique - I suspect that this is due to the NULL values in the NAPTAN_ATCO parameter. Issue to be interrogated this week.